use getStaticProps
You can use this function exactly as stated to have hidden code to the public so that when the client side is rendered, the function only would get the static props needed and have it rendered to the client beforehand. The code inside the function is static items. This is what's being displayed on the screen