Create second route with dynamic staticProps
The second route contains data about that product like title and description only. This uses static props too with the same method as before using fs except this time having the find logic inside the file unlike with the previous project it was in the data file. The get static props here returns loadedProduct inside the file with product as the value and so no information is passed down using react context