Fallback feature
Setting fallback to false means the params has to be in the paths array. Setting it to true means it doesnt need to be in there but a client cannot navigate to there by using the search at the top. In this case, you can conditionally render a loading if loadingProduct isnt defined. However, having the string of blocking as the value for fallback means dealing with both issues but this will take longer since nothing is pre-rendered and the page will render on time